+++ pychron/processing/tasks/figures/actions.py
@@ -53,23 +53,23 @@ class OpenFigureAction(TaskAction):
 
 class NewXYScatterAction(TaskAction):
     name = 'New XY Scatter'
-    method = 'tb_xy_scatter'
+    method = 'new_xy_scatter'
 
 
 class NewIsochronAction(TaskAction):
-    name = 'New Isochron'
-    method = 'tb_new_isochron'
+    name = 'New Inv. Isochron'
+    method = 'new_inverse_isochron'
 
 
 class NewIdeogramAction(TaskAction):
     name = 'New Ideogram'
-    method = 'tb_new_ideogram'
+    method = 'new_ideogram'
     image = icon('histogram')
 
 
 class NewSpectrumAction(TaskAction):
     name = 'New Spectrum'
-    method = 'tb_new_spectrum'
+    method = 'new_spectrum'
     image = icon('chart_curve')
 
 
@@ -79,6 +79,13 @@ class RefreshActiveEditorAction(TaskAction):
     image = icon('refresh')
     # accelerator = 'Ctrl+Shift+R'
     id = 'pychron.refresh_plot'
+
+
+class NewTableAction(TaskAction):
+    name = 'New Table'
+    method = 'new_table'
+    image = icon('table')
+
 # ============= EOF =============================================
 # class AppendSpectrumAction(TaskAction):
 #     name = 'Append Spectrum'